Here we are trying to identify the fraud transactions on credit cards using machinie Learning algorithm of logistic regression. The datasets contains transactions made by credit cards in September 2013 by european cardholders. This dataset presents transactions that occurred in two days, where we have 492 frauds out of 284,807 transactions. The dataset is highly unbalanced, the positive class (frauds) account for 0.172% of all transactions. We were able to successfully predit the 63% of  the fraud transactions using our algorithm and almost 100% of valid transactions.
The datasets contains transactions made by credit cards in September 2013 by european cardholders. This dataset presents transactions that occurred in two days, where we have 492 frauds out of 284,807 transactions. The dataset is highly unbalanced, the positive class (frauds) account for 0.172% of all transactions.

It contains only numerical input variables which are the result of a PCA transformation. Unfortunately, due to confidentiality issues, we cannot provide the original features and more background information about the data. Features V1, V2, ... V28 are the principal components obtained with PCA, the only features which have not been transformed with PCA are 'Time' and 'Amount'. Feature 'Time' contains the seconds elapsed between each transaction and the first transaction in the dataset. The feature 'Amount' is the transaction Amount, this feature can be used for example-dependant cost-senstive learning. Feature 'Class' is the response variable and it takes value 1 in case of fraud and 0 otherwise.

Use Logistic Regression

Train on a portion of the dataset
Test the trained model against the remainder of the dataset
Accuracy can be determined because the dataset is labeled (i.e., this uses supervised learning)
Result:

In this kind of problem, instead of acheving an oveall accuracy, we would like to know how many fraud transactions were correctly identified and how many were missed.
Using Logistic Regression for prediction of fraud trannsaction, we were able to acheve a 63% accuracy in identifying the number of frauds.